Normalize data using common pools

Normalize nanostring gene expression using common pools between two CodeSets.

Usage

normalize_pools(
  x,
  x_pools,
  ref_pools,
  p = 3,
  weigh = TRUE,
  same_codeset = FALSE
)

Arguments

x

target data

x_pools

target pool samples

ref_pools

reference pool samples

p

number of pool sample sets. Defaults to 3.

weigh

logical; if TRUE, the average expression in x_pools is reweighed by the distribution of the p pool sample sets in ref_pools.

same_codeset

logical; if TRUE, the target pool samples are collected from the same codset as the reference pool samples. This means the pool numbers from the file names use an older format where two-digit pools still pertain to pools 1, 2, or 3.

Value

normalized gene expression

Details

The target and reference expression samples, as well the target and reference pool samples all need to be specified. We recommend reweighing the target pool samples when calculating the average expression by the distribution of reference pools.

Author

Derek Chiu