Normalize data using random reference samples

Normalize nanostring gene expression using randomly chosen samples for the reference-based approach for batch adjustment.

normalize_random(x, ref, n = 1, strata = NULL, seed = NULL)

Arguments

x

target data

ref

reference data

n

number of random reference samples to select for normalization

strata

a grouping variable for stratified random sampling. If strata has k levels, then n * k random samples are selected.

seed

random seed for reproducibility

Value

normalized gene expression

Details

The number of randomly chosen numbers can be selected, and optionally a strata can be specified such that n reference samples are selected from each level (like a stratified bootstrap). In relation to the reference method, the random samples removed from ref form R1, the random samples removed from x form R2, and the remaining samples from x form Y. See refMethod() for details.

In subsequent analyses, we refer to a method using normalize_random(n) as the "Random n" method.

Author

Derek Chiu