Read NanoString RCC files

Read RCC files and extract count and attribute data. Use read_rcc() for multiple files, and use the parse_*() functions for single files.

Usage

read_rcc(path = ".")

parse_counts(file)

parse_attributes(file)

Arguments

path

directory path for multiple RCC files

file

RCC file name

Value

read_rcc() reads in a directory of RCC files and outputs a list with two elements:

raw: A tibble of parsed counts for multiple RCC files created by calling parse_counts() on each sample. Columns include "Code.Class", "Name", "Accession", and a column for each sample ID. There is one row per gene.

exp: A tibble of parsed attributes for multiple RCC files created by calling parse_attributes() on each sample. Columns include "File.Name" (sample ID), "geneRLF", "nanostring.date", "cartridgeID", "lane.number", fov.count", "fov.counted", "binding.density". There is one row per sample.

parse_counts() reads a single RCC file and returns a tibble of parsed counts.

parse_attributes() reads a single RCC file and returns a list of parsed attributes.

Details

RCC files for a sample are direct outputs from NanoString runs. We can extract counts for each gene in a sample. Sample attributes include sample ID, GeneRLF, date, cartridge ID, lane number, Fov count, Fov counted, and binding density. read_rcc() merges both count and attribute data across samples.

If path points to a zipped RCC file with multiple samples, the zip file is uncompressed and a directory of RCC sample files is created with the same name. Only file extensions ".RCC" or ".rcc" are allowed.

Author

Derek Chiu

Examples

rcc_file <- system.file("extdata", "example.RCC", package = "nanostringr")
parse_counts(rcc_file)
#> # A tibble: 354 × 4
#>    Code.Class Name    Accession      `TAOC00241-N1`
#>    <chr>      <chr>   <chr>                   <int>
#>  1 Endogenous COL1A1  NM_000088.3             47179
#>  2 Endogenous ADAM12  NM_003474.4               158
#>  3 Endogenous BRCA2   NM_000059.3                32
#>  4 Endogenous GTF2H5  NM_207118.2               117
#>  5 Endogenous SLC7A11 NM_014331.3                33
#>  6 Endogenous HIF1A   NM_001530.2               691
#>  7 Endogenous DCN     NM_001920.3              2431
#>  8 Endogenous SVIL    NM_003174.2               335
#>  9 Endogenous HBB     NM_000518.4               292
#> 10 Endogenous MRPS27  NM_001286748.1             91
#> # ℹ 344 more rows
parse_attributes(rcc_file)
#> $File.Name
#> [1] "TAOC00241-N1"
#> 
#> $geneRLF
#> [1] "OTTA2017_C6082"
#> 
#> $nanostring.date
#> [1] "2018-07-24"
#> 
#> $cartridgeID
#> [1] "OTTA2-108-SYD-20H"
#> 
#> $lane.number
#> [1] "4"
#> 
#> $fov.count
#> [1] 555
#> 
#> $fov.counted
#> [1] 542
#> 
#> $binding.density
#> [1] 0.27
#> 
read_rcc(dirname(rcc_file))
#> $raw
#>       Code.Class         Name         Accession example
#> 1     Endogenous       COL1A1       NM_000088.3   47179
#> 2     Endogenous       ADAM12       NM_003474.4     158
#> 3     Endogenous        BRCA2       NM_000059.3      32
#> 4     Endogenous       GTF2H5       NM_207118.2     117
#> 5     Endogenous      SLC7A11       NM_014331.3      33
#> 6     Endogenous        HIF1A       NM_001530.2     691
#> 7     Endogenous          DCN       NM_001920.3    2431
#> 8     Endogenous         SVIL       NM_003174.2     335
#> 9     Endogenous          HBB       NM_000518.4     292
#> 10    Endogenous       MRPS27    NM_001286748.1      91
#> 11    Endogenous        CXCR3       NM_001504.1      32
#> 12    Endogenous        FGFR1       NM_015850.2     690
#> 13    Endogenous        ADRM1    NM_001281437.1     297
#> 14    Endogenous        PSMB8       NM_004159.4     152
#> 15    Endogenous        NLRP2       NM_017852.1     240
#> 16    Endogenous        MATN3       NM_002381.4      97
#> 17    Endogenous          FAP       NM_004460.2     293
#> 18    Endogenous        PCDH9       NM_020403.3      16
#> 19    Endogenous       MAGEA4    NM_001011548.1       4
#> 20    Endogenous         MYCN       NM_005378.4      34
#> 21    Endogenous        CDKN3       NM_005192.3      32
#> 22    Endogenous         CILP       NM_003613.3     111
#> 23    Endogenous         PGRA       NM_000926.4      30
#> 24    Endogenous        OR1G1       NM_003555.1      10
#> 25    Endogenous         PCK2       NM_004563.2      71
#> 26    Endogenous         PGRB       NM_000926.4      13
#> 27    Endogenous       UBE2L6       NM_004223.3     298
#> 28    Endogenous        KLHL7    NM_001172428.1      62
#> 29    Endogenous         EYA2       NM_005244.4     211
#> 30    Endogenous       CXCL10       NM_001565.1      46
#> 31    Endogenous       COL3A1       NM_000090.3   40585
#> 32    Endogenous       OLFML3       NM_020190.2     708
#> 33    Endogenous       Notch3       NM_000435.2     609
#> 34    Endogenous         ARNT       NM_001668.2     128
#> 35    Endogenous        ANXA4       NM_001153.2     196
#> 36    Endogenous        RAD50       NM_005732.2      72
#> 37    Endogenous       IGFBP2       NM_000597.2    3549
#> 38    Endogenous       TSPAN8       NM_004616.2       5
#> 39    Endogenous       SEMA6A       NM_020796.3     104
#> 40    Endogenous         IGKC        AJ294735.1    3161
#> 41    Endogenous        ENOX1    NM_001242863.1      50
#> 42    Endogenous          SYK       NM_003177.5      66
#> 43    Endogenous         ROR2       NM_004560.2     232
#> 44    Endogenous         GMNN       NM_015895.3      44
#> 45    Endogenous         CCL5       NM_002985.2     182
#> 46    Endogenous        KLK10    NM_001077500.1      86
#> 47    Endogenous        CD151       NM_139029.1     255
#> 48    Endogenous         IL6R       NM_000565.2     146
#> 49    Endogenous        SNAI2       NM_003068.3     323
#> 50    Endogenous      GUCY1A2       NM_000855.1      57
#> 51    Endogenous          CD2       NM_001767.2      38
#> 52    Endogenous        ABCF1       NM_001090.2     337
#> 53    Endogenous        RAD52       NM_134424.3      11
#> 54    Endogenous         ARAF       NM_001654.3     224
#> 55    Endogenous       CX3CR1       NM_001337.3      23
#> 56    Endogenous        BAALC       NM_024812.2      28
#> 57    Endogenous        TUBB6       NM_032525.1     809
#> 58    Endogenous        NTRK2    NM_001007097.1      41
#> 59    Endogenous        OBFC1    XM_006717976.3      19
#> 60    Endogenous          MYC       NM_002467.3     809
#> 61    Endogenous        RASA1       NM_002890.1     211
#> 62    Endogenous       FAM58A       NM_152274.3      85
#> 63    Endogenous       ASRGL1       NM_025080.3      33
#> 64    Endogenous        CPNE1    NM_001198863.1     200
#> 65    Endogenous        FANCF       NM_022725.2      74
#> 66    Endogenous        DUSP4       NM_057158.2      96
#> 67    Endogenous     SERPINE1       NM_000602.4    1796
#> 68    Endogenous      ALDH5A1       NM_001080.3      59
#> 69    Endogenous        CDH11       NM_001797.2     934
#> 70    Endogenous      SLC15A3       NM_016582.1     450
#> 71    Endogenous         RFC2       NM_181471.1      56
#> 72    Endogenous        MFAP5       NM_003480.2     104
#> 73    Endogenous       CXCL12       NM_199168.3    1593
#> 74    Endogenous        WDR91       NM_014149.3     109
#> 75    Endogenous         IRF1       NM_002198.1     168
#> 76    Endogenous         ZEB1    NM_001128128.1     170
#> 77    Endogenous        RAD17       NM_133338.1      17
#> 78    Endogenous        FOXJ1       NM_001454.3      34
#> 79    Endogenous        STK39       NM_013233.2      58
#> 80    Endogenous       CXCL11       NM_005409.3      50
#> 81    Endogenous        CD302       NM_014880.4      78
#> 82    Endogenous       FCHSD2       NM_014824.2     104
#> 83    Endogenous       SUPT6H       NM_003170.3     248
#> 84    Endogenous         PAX8       NM_013953.3     523
#> 85    Endogenous      B4GALT5       NM_004776.3     395
#> 86    Endogenous       CRABP1       NM_004378.2       5
#> 87    Endogenous          NBN       NM_002485.4     169
#> 88    Endogenous       DNAJC9       NM_015190.3      79
#> 89    Endogenous       ZNF165       NM_003447.3      63
#> 90    Endogenous        PD-L1       NM_014143.3      41
#> 91    Endogenous        GFRA1       NM_005264.4      24
#> 92    Endogenous      SLC4A11    NM_001174090.1      80
#> 93    Endogenous        DGCR2    NM_001173534.1     107
#> 94    Endogenous        PTH2R       NM_005048.2      30
#> 95    Endogenous       FCER1G       NM_004106.1     266
#> 96    Endogenous      SMARCA4       NM_003072.3     146
#> 97    Endogenous        ABCC2       NM_000392.3       7
#> 98    Endogenous        UCHL1       NM_004181.4      28
#> 99    Endogenous        RCCD1    NM_001017919.1      48
#> 100   Endogenous          RB1       NM_000321.1      95
#> 101   Endogenous        DPPA2       NM_138815.2       6
#> 102   Endogenous      TMEM45A       NM_018004.1     168
#> 103   Endogenous         FEN1       NM_004111.4      18
#> 104   Endogenous         EXO1       NM_003686.3      26
#> 105   Endogenous        ABCE1    NM_001040876.1     138
#> 106   Endogenous    HIST1H2AM       NM_003514.2     426
#> 107   Endogenous         MITF       NM_000248.2      37
#> 108   Endogenous       PPP2R4       NM_021131.4      80
#> 109   Endogenous       SLAMF8       NM_020125.2      17
#> 110   Endogenous         CDH1       NM_004360.2     295
#> 111   Endogenous       RAD23A       NM_005053.2     165
#> 112   Endogenous          LUM       NM_002345.3    5282
#> 113   Endogenous     SERPINF1       NM_002615.4    1050
#> 114   Endogenous         BST2       NM_004335.2     137
#> 115   Endogenous         ULK1       NM_003565.1     118
#> 116   Endogenous          AXL       NM_021913.2     259
#> 117   Endogenous        SALL2       NM_005407.1      64
#> 118   Endogenous         ESR2    NM_001040275.1      17
#> 119   Endogenous        THBS1       NM_003246.2    4480
#> 120   Endogenous        MMP11       NM_005940.3    1428
#> 121   Endogenous        BRCA1       NM_007294.3       8
#> 122   Endogenous         IDO1       NM_002164.3      54
#> 123   Endogenous     ABRAXAS2       NM_032182.3      44
#> 124   Endogenous         TLR4       NR_024168.1      75
#> 125   Endogenous    HIST1H2BD       NM_021063.2     188
#> 126   Endogenous        MUC16       NM_024690.2     126
#> 127   Endogenous      SCGB1D2       NM_006551.3       6
#> 128   Endogenous          MAK       NM_005906.3      29
#> 129   Endogenous       Drosha    NM_001100412.1     122
#> 130   Endogenous         OPA1       NM_130837.1     240
#> 131   Endogenous         CAV1       NM_001753.3     112
#> 132   Endogenous        CNOT4    NM_001190848.1     106
#> 133   Endogenous        LAMC2       NM_005562.2     715
#> 134   Endogenous          SRI       NM_003130.2     388
#> 135   Endogenous         FBN1       NM_000138.3     359
#> 136   Endogenous        CTLA4       NM_005214.3      34
#> 137   Endogenous          TNC       NM_002160.3     113
#> 138   Endogenous      C9orf82    NM_001167575.1      30
#> 139   Endogenous        VCAM1       NM_001078.3     454
#> 140   Endogenous       EFEMP1       NM_004105.3     112
#> 141   Endogenous         PAX2       NM_000278.3      21
#> 142   Endogenous        ENPP1       NM_006208.1     154
#> 143   Endogenous       AKT1S1       NM_032375.3      77
#> 144   Endogenous       CXCL14       NM_004887.4     400
#> 145   Endogenous         DAB2       NM_001343.2     473
#> 146   Endogenous         GJB1       NM_000166.5       6
#> 147   Endogenous         VCAN       NM_004385.3     488
#> 148   Endogenous       POLR2E       NM_002695.3      39
#> 149   Endogenous         PD-1       NM_005018.1      11
#> 150   Endogenous      RHOBTB3       NM_014899.3     140
#> 151   Endogenous      COL10A1       NM_000493.3    1047
#> 152   Endogenous        HMGA2       NM_003484.1     157
#> 153   Endogenous       MAP2K4       NM_003010.2     127
#> 154   Endogenous         CD3D       NM_000732.4      44
#> 155   Endogenous         USP8    NM_001128610.1     238
#> 156   Endogenous        APBB2       NM_173075.4     242
#> 157   Endogenous         CD74    NM_001025158.1    4406
#> 158   Endogenous         IGHM ENST00000390559.1     118
#> 159   Endogenous          SMO       NM_005631.3      43
#> 160   Endogenous        FABP4       NM_001442.2     360
#> 161   Endogenous         IL22       NM_020525.4      10
#> 162   Endogenous        XRCC6       NM_001469.3     256
#> 163   Endogenous         TAP1       NM_000593.5      77
#> 164   Endogenous        PTTG1       NM_004219.2      48
#> 165   Endogenous         TSHR    NM_001018036.2       4
#> 166   Endogenous        ZFHX4       NM_024721.4      76
#> 167   Endogenous        CAPN9       NM_006615.2       7
#> 168   Endogenous        INHBA       NM_002192.2     758
#> 169   Endogenous       ANKRD1       NM_014391.2      26
#> 170   Endogenous      CTNNBL1       NM_030877.3      36
#> 171   Endogenous        TAGLN       NM_003186.3    3567
#> 172   Endogenous      ALDH1A1       NM_000689.3      76
#> 173   Endogenous        NUCB2       NM_005013.2     151
#> 174   Endogenous        FZD10       NM_007197.2      32
#> 175   Endogenous       SPTLC2       NM_004863.3      72
#> 176   Endogenous         CD38       NM_001775.2      21
#> 177   Endogenous        SPARC       NM_003118.2   27155
#> 178   Endogenous         PLK2       NM_006622.2     357
#> 179   Endogenous         TBX2       NM_005994.3      85
#> 180   Endogenous         ESR1       NM_000125.2     326
#> 181   Endogenous        ABCC1       NM_004996.3     159
#> 182   Endogenous         HIC5    NM_001042454.1     475
#> 183   Endogenous         COBL       NM_015198.3      45
#> 184   Endogenous         MCM3       NM_002388.3     188
#> 185   Endogenous        ITGB5       NM_002213.3     559
#> 186   Endogenous      COL11A1       NM_001854.3    1420
#> 187   Endogenous      RARRES1       NM_002888.2     201
#> 188   Endogenous         PCNA       NM_002592.2     122
#> 189   Endogenous           AR    NM_001011645.1      56
#> 190   Endogenous         MEST       NM_177525.1     189
#> 191   Endogenous          EZR       NM_003379.4     960
#> 192   Endogenous        TUBB2       NM_001069.2     223
#> 193   Endogenous        PARP1       NM_001618.3     170
#> 194   Endogenous       LRRC15    NM_001135057.2     510
#> 195   Endogenous       COL1A2       NM_000089.3   27607
#> 196   Endogenous          LOX       NM_002317.4     428
#> 197   Endogenous         CD68       NM_001251.2     274
#> 198   Endogenous         CD8A       NM_001768.5     143
#> 199   Endogenous       CDKN2A       NM_000077.4      81
#> 200   Endogenous        APPL2       NM_018171.3      49
#> 201   Endogenous        PRAME       NM_006115.4     184
#> 202   Endogenous         KRT6       NM_005554.3      15
#> 203   Endogenous        DUSP1       NM_004417.2    5327
#> 204   Endogenous         IGF2       NM_000612.4     243
#> 205   Endogenous        NDRG3       NM_032013.3      81
#> 206   Endogenous      RARRES2       NM_002889.3    3655
#> 207   Endogenous        THBS2       NM_003247.2    3168
#> 208   Endogenous       ZNF423       NM_015069.2      46
#> 209   Endogenous        DLEC1       NM_007335.2      22
#> 210   Endogenous      TPD52L1       NM_003287.2      54
#> 211   Endogenous         OASL       NM_198213.1      25
#> 212   Endogenous          APC       NM_000038.3      91
#> 213   Endogenous         WWP1       NM_007013.3     170
#> 214   Endogenous         TP53       NM_000546.2     172
#> 215   Endogenous       PIK3CA       NM_006218.2     288
#> 216   Endogenous        PTGIS       NM_000961.3     704
#> 217   Endogenous        POSTN       NM_006475.2     490
#> 218   Endogenous        TIMP3       NM_000362.4    3969
#> 219   Endogenous       RAD23B       NM_002874.3     204
#> 220   Endogenous        KIFC1       NM_002263.3      99
#> 221   Endogenous       BNIP3L       NM_004331.2     275
#> 222   Endogenous       RAD51B       NM_002877.4      26
#> 223   Endogenous       GALNT6       NM_007210.3     106
#> 224   Endogenous        NUAK1       NM_014840.2     136
#> 225   Endogenous RP11-89M16.1       NR_121672.1      34
#> 226   Endogenous       ATP5A1    NM_001001937.1     475
#> 227   Endogenous          WT1       NM_000378.3     538
#> 228   Endogenous      TMEM158       NM_015444.2    4140
#> 229   Endogenous        SFRP4       NM_003014.2    5043
#> 230   Endogenous       SNRPA1       NM_003090.2     168
#> 231   Endogenous        CXCR4       NM_003467.2    1478
#> 232   Endogenous        YWHAB       NM_003404.4     858
#> 233   Endogenous      EPB41L3       NM_012307.2     150
#> 234   Endogenous         FGF1       NM_033137.1      52
#> 235   Endogenous      SCGB2A1       NM_002407.2      23
#> 236   Endogenous         MSLN       NM_013404.3    1501
#> 237   Endogenous      FOXRED2    NM_001102371.1      57
#> 238   Endogenous        STK16       NM_003691.2      88
#> 239   Endogenous        TESK1       NM_006285.2      29
#> 240   Endogenous        DUSP6       NM_001946.2     209
#> 241   Endogenous         CTSK       NM_000396.2    2363
#> 242   Endogenous       SLAMF7       NM_021181.3      90
#> 243   Endogenous         PTK2       NM_005607.4     243
#> 244   Endogenous        PARP4       NM_006437.3     289
#> 245   Endogenous        CXCL9       NM_002416.1     233
#> 246   Endogenous        CYTIP       NM_004288.4     100
#> 247   Endogenous       MRE11A       NM_005591.3      81
#> 248   Endogenous          ESD       NM_001984.1     193
#> 249   Endogenous       PDGFRB       NM_002609.3     314
#> 250   Endogenous         UCP2       NM_003355.2      61
#> 251   Endogenous       MINPP1    NM_001178117.1      39
#> 252   Endogenous        NLRP7       NM_139176.3      33
#> 253   Endogenous          FN1       NM_002026.2    7163
#> 254   Endogenous          IGJ       NM_144646.3      84
#> 255   Endogenous        AEBP1       NM_001129.3    4719
#> 256   Endogenous         PGCP       NM_016134.2     160
#> 257   Endogenous       IGFBP4       NM_001552.2    3068
#> 258   Endogenous       TCF7L1       NM_031283.1      43
#> 259   Endogenous        CCNE1       NM_001238.1      63
#> 260   Endogenous        CFLAR    NM_001127183.1      90
#> 261   Endogenous        NEAT1       NR_028272.1    9235
#> 262   Endogenous         GUSB       NM_000181.3     224
#> 263   Endogenous       CRABP2       NM_001878.2     858
#> 264   Endogenous         TUG1       NR_002323.1     174
#> 265   Endogenous        CSF1R       NM_005211.2      33
#> 266   Endogenous       FBX038       NM_030793.4     124
#> 267   Endogenous       SAC3D1       NM_013299.3      32
#> 268   Endogenous        KIF1A       NM_004321.4     157
#> 269   Endogenous        ADH1B       NM_000668.4     218
#> 270   Endogenous        VSIG4    NM_001100431.1     172
#> 271   Endogenous       L3MBTL       NM_015478.6      41
#> 272   Endogenous        TGFB2       NM_003238.2      56
#> 273   Endogenous        CPEB1       NM_030594.3      20
#> 274   Endogenous        ZFAS1       NR_003605.1     830
#> 275   Endogenous        LUZP1       NM_033631.3     114
#> 276   Endogenous       MAP4K3       NM_003618.2     194
#> 277   Endogenous         FLNC    NM_001127487.1     273
#> 278   Endogenous         CLTA    NM_001184762.1     293
#> 279   Endogenous          NF1       NM_000267.2     181
#> 280   Endogenous         CDK6       NM_001259.6     178
#> 281   Endogenous        HNF1B       NM_000458.1      16
#> 282   Endogenous     ADAMDEC1       NM_014479.2      14
#> 283   Endogenous        MYOD1       NM_002478.4       5
#> 284   Endogenous       TRIM27       NM_006510.4     109
#> 285   Endogenous        ABCB1       NM_000927.3      54
#> 286   Endogenous        WDR77       NM_024102.2     283
#> 287   Endogenous     HSP90AA1    NM_001017963.2    1105
#> 288   Endogenous        CCL20       NM_004591.1      84
#> 289   Endogenous     C19orf12       NM_031448.3      63
#> 290   Endogenous       CTHRC1       NM_138455.2    2105
#> 291   Endogenous         ROR1    NM_001083592.1      85
#> 292   Endogenous         MDM2       NM_032739.1     164
#> 293   Endogenous        NUAK2       NM_030952.1     408
#> 294   Endogenous     CRISPLD2       NM_031476.3    1332
#> 295   Endogenous        NSDHL    NM_001129765.1      66
#> 296   Endogenous       COL5A2       NM_000393.3    2686
#> 297   Endogenous         GMPR       NM_006877.3      91
#> 298   Endogenous        FOLR1       NM_000802.2     160
#> 299   Endogenous       CYP4B1       NM_000779.3      83
#> 300   Endogenous        ABCC3    NM_001144070.1      27
#> 301   Endogenous        SHPRH    NM_001042683.2      77
#> 302   Endogenous       RNASEL       NM_021133.3      58
#> 303   Endogenous        LPAR3       NM_012152.1     337
#> 304   Endogenous         IGF1       NM_000618.3     793
#> 305   Endogenous         POLE       NM_006231.3      31
#> 306   Endogenous         E2F6       NM_198325.1      96
#> 307   Endogenous         PTEN       NM_000314.3     506
#> 308   Endogenous         MDM4    NM_001204172.1     136
#> 309   Endogenous       LGALS1       NM_002305.3    2941
#> 310   Endogenous       COL5A1       NM_000093.3    2297
#> 311   Endogenous       POLR2I       NM_006233.4     213
#> 312   Endogenous        RBMS3    NM_001003792.2      33
#> 313   Endogenous         ZIP4       NM_017767.2      49
#> 314   Endogenous        ABHD1       NM_032604.3      10
#> 315   Endogenous         UQCC       NM_199487.2      72
#> 316   Endogenous         MMP7       NM_002423.3    3811
#> 317   Endogenous        Dicer       NM_177438.2     186
#> 318   Endogenous        PRKCQ       NM_006257.2      48
#> 319   Endogenous        FOXP3       NM_014009.3      18
#> 320   Endogenous         MMP2       NM_004530.2     909
#> 321   Endogenous        EDNRA       NM_001957.1     106
#> 322   Endogenous        KDM5D    NM_001146705.1       4
#> 323   Endogenous         CD3e       NM_000733.2      33
#> 324   Endogenous       RAD51C       NM_002876.2      67
#> 325   Endogenous     PDZK1IP1       NM_005764.3      23
#> 326   Endogenous          PI3       NM_002638.3      18
#> 327   Endogenous         PVT1       NR_003367.1      26
#> 328   Endogenous         CD27       NM_001242.4      22
#> 329   Endogenous       IGFBP6       NM_002178.2     110
#> 330   Endogenous          MAL       NM_002371.2      55
#> 331   Endogenous       PTGER3       NM_000957.2     125
#> 332   Endogenous       HIGD1B       NM_016438.2       9
#> 333   Endogenous        SOX17       NM_022454.3     294
#> 334   Endogenous        GFPT2       NM_005110.2      73
#> 335   Endogenous        PRKCI       NM_002740.5    1655
#> 336     Positive     POS_C(8)      ERCC_00002.1    4307
#> 337     Positive   POS_A(128)      ERCC_00117.1   56906
#> 338     Positive POS_F(0.125)      ERCC_00034.1      94
#> 339     Positive     POS_D(2)      ERCC_00092.1    1064
#> 340     Positive    POS_B(32)      ERCC_00112.1   14747
#> 341     Positive   POS_E(0.5)      ERCC_00035.1     220
#> 342     Negative     NEG_C(0)      ERCC_00019.1       3
#> 343     Negative     NEG_D(0)      ERCC_00076.1       4
#> 344     Negative     NEG_E(0)      ERCC_00098.1      21
#> 345     Negative     NEG_A(0)      ERCC_00096.1      14
#> 346     Negative     NEG_H(0)      ERCC_00154.1       0
#> 347     Negative     NEG_G(0)      ERCC_00144.1       2
#> 348     Negative     NEG_F(0)      ERCC_00126.1      11
#> 349     Negative     NEG_B(0)      ERCC_00041.1       5
#> 350 Housekeeping       POLR1B       NM_019014.3      80
#> 351 Housekeeping         PGK1       NM_000291.2    1445
#> 352 Housekeeping        RPL19       NM_000981.3    9440
#> 353 Housekeeping         ACTB       NM_001101.2   11985
#> 354 Housekeeping         SDHA       NM_004168.1     110
#> 
#> $exp
#>    sample        geneRLF nanostring.date       cartridgeID lane.number
#> 1 example OTTA2017_C6082      2018-07-24 OTTA2-108-SYD-20H           4
#>   fov.count fov.counted binding.density
#> 1       555         542            0.27
#>